Include Enumerable in ROM::Relation::Materializable
This means that `#each` doesn't work differently to other `Enumerable`
methods like `#map`, which are currently forwarded to the relation.

The current behavior is very surprising: `#each` will yield model
objects, while `#map` and other `Enumerable` methods will yield hashes.

Since `Enumerable` provides a `#find` method, a spec that relied on
`Materializable` not having a `#find` had to be updated to use a
different non-existent method.